Reenable interrupts after returning from 25h/26h.
In attempting to debug the game freezing up on multiplayer startup, it
looks like it was waiting forever in a `while ( _dkiosts() );` line
blocking for a timer interrupt with interrupts disabled.
I think this is because the FLAGS register was just being discarded by
the Sopwith code and not restored (since we aren't doing an iret).