Rework how loopback sends work.
The previous approach copied the entire send window to the receive
window. This doesn't work because both the send and receive code need to
be operating on the same data structure; if the command was popped from
the receive window it was still sitting in the send window. Instead,
special-case the local player so that we write directly into its
receive window.